requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
telethon==1.38.1
PyYAML==6.0.2
